requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
argcomplete==3.6.1
asgiref==3.8.1
astunparse==1.6.3
attr==0.3.1
attrs==25.3.0
azure-core==1.32.0
azure-storage-blob==12.25.0
black==25.1.0
bleach==5.0.1
blinker==1.9.0
boto==2.49.0
boto3==1.37.19
botocore==1.37.19
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.7.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
datamodel-code-generator==0.26.1
DAWG2-Python==0.9.0
defusedxml==0.7.1
Deprecated==1.2.18
distro==1.9.0
Django==5.1.7
django-annoying==0.10.6
django-cors-headers==3.6.0
django-csp==3.7
django-debug-toolbar==3.2.1
django-environ==0.10.0
django-extensions==3.2.3
django-filter==24.3
django-migration-linter==5.1.0
django-model-utils==4.1.1
django-ranged-fileresponse==0.1.2
django-rq==2.10.3
django-storages==1.12.3
django-user-agents==0.4.0
djangorestframework==3.15.2
dnspython==2.7.0
drf-dynamic-fields==0.3.0
drf-flex-fields==0.9.5
drf-generators==0.3.0
email_validator==2.2.0
expiringdict==1.2.2
Faker==37.1.0
fasttext-wheel==0.9.2
filelock==3.17.0
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.0
gast==0.6.0
genson==1.3.0
google-api-core==2.24.2
google-auth==2.38.0
google-cloud-appengine-logging==1.6.1
google-cloud-audit-log==0.3.2
google-cloud-core==2.4.3
google-cloud-logging==3.11.4
google-cloud-storage==2.19.0
google-crc32c==1.7.1rc1
google-pasta==0.2.0
google-resumable-media==2.7.2
googleapis-common-protos==1.69.2
grpc-google-iam-v1==0.14.2
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.3
humanfriendly==10.0
humansignal-drf-yasg==1.21.10.post1
idna==3.10
ijson==3.3.0
imageio==2.37.0
importlib_metadata==8.6.1
inaSpeechSegmenter==0.7.12
inflect==5.6.2
inflection==0.5.1
isodate==0.7.2
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsf==0.11.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keras==3.9.0
kiwisolver==1.4.8
label-studio==1.16.0
label-studio-sdk==1.0.10
launchdarkly-server-sdk==8.2.1
lazy_loader==0.4
libclang==18.1.1
llvmlite==0.44.0
lockfile==0.12.2
lxml==5.3.1
lxml_html_clean==0.4.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
ml-dtypes==0.4.1
more-itertools==10.6.0
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.30.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
noisereduce==3.0.3
NRCLex==3.0.0
numba==0.61.0
numpy==1.26.4
onnxruntime-gpu==1.20.1
openai==1.68.2
openai-whisper==20240930
opentelemetry-api==1.31.1
opt_einsum==3.4.0
optree==0.14.1
ordered-set==4.0.2
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.7
plotly==6.0.0
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyannote.core==5.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybind11==2.13.6
pyboxen==1.3.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pymorphy3==2.0.3
pymorphy3-dicts-ru==2.4.417150.4580142
pyparsing==3.2.1
pyreadline3==3.5.4
pyRFC3339==2.0.1
Pyro4==4.82
pytextgrid==0.1.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.4
pytz==2022.7.1
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-mock==1.12.1
retrying==1.3.4
rich==13.9.4
rpds-py==0.23.1
rq==1.16.2
rsa==4.9
rstr==3.2.2
rules==3.4
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
semver==3.0.4
sentry-sdk==2.24.1
serpent==1.41
setuptools==75.8.2
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.13.1
sqlparse==0.5.3
srt==3.5.3
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-intel==2.17.0
termcolor==2.5.0
textblob==0.19.0
tf-keras==2.15.0
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.21.0
toml==0.10.2
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
tzdata==2025.1
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
ujson==5.10.0
uritemplate==4.1.1
urllib3==1.26.20
user-agents==2.2.0
vosk==0.3.45
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.0.6
wheel==0.40.0
wrapt==1.17.2
xmljson==0.2.1
zipp==3.21.0
